Move queryVersions and getPackageManager to /src/lib
